+++ NEWS.md
@@ -10,8 +10,8 @@
 
 ## Changes in version 1.1.3
 
-* Fixed a bug affecting coverage related to gaps in the alignment in MapAssessmentData
-* Used exponential moving averages to filter out bad alignmnets in MapAssessmentData
+* Fixed a bug affecting coverage related to gaps in alignments in MapAssessmentData
+* Used exponential moving averages to filter out bad alignment regions in MapAssessmentData
 
 # AssessORF 0.99